#187 fix svg width/height issues

+++ blaise-svg/src/main/java/com/googlecode/blaisemath/graphics/svg/SVGGraphic.java
@@ -97,8 +97,8 @@ public class SVGGraphic extends GraphicComposite<Graphics2D> {
         primitiveElement = nue;
         
         if (element instanceof SVGRoot) {
-            int wid = ((SVGRoot) element).getWidth();
-            int ht = ((SVGRoot) element).getHeight();
+            double wid = ((SVGRoot) element).getWidth();
+            double ht = ((SVGRoot) element).getHeight();
             this.graphicBounds = new Rectangle2D.Double(0, 0, wid, ht);
         }
     }
